requirements.txt
pygame
pyserial
more-itertools
pylsl
numpy
scipy
sklearn
matplotlib
optuna
